more refactoring on repo and fork stuff
